replace global interrupt map with per-node maps
Local maps are more legible, since it's easier to see where are
the node's interrupts wired to.

Also, local maps eliminate the need for globally unique interrupt IDs,
whose meaning is limited to this device tree, without a meaning in real
HW. Instead, we only have interrupt IDs unique to the scope of each
interrupt controller (note that we don't yet take advantage of that
scoping and currently interrupt controller inputs that are used
are index by a global ID -- to be changed shortly).