hpps smmu: fix address-cells in incoming ports
Was broken in 07367a7a.

The #address-cells property that applies to a port is the one
that applies to the port node, not the one that applies to the
smmu node. In this case, that's 1 cell, not 2 cells.

This fixes Ethernet connectivity (and unmapped memory accesses
when going through SMMU) and DMA tests.